Logo: the lighter linked loops; tidier Today scores; short month
The mark and home-screen icons switch to the thinner, more open pair of
links. The scoreboard drops the "+N today" line under each name, and the
date at the top uses the short month (Sept).

+++ src/start/Home.tsx
@@ -133,7 +133,7 @@ function ProfileButton() {
 function Today({ onPick }: { onPick: (g: Game) => void }) {
   const board = useBoard()
   const paired = board.status.kind === 'ready' && board.status.data.state === 'paired' ? board.status.data : null
-  const date = new Date().toLocaleDateString(undefined, { weekday: 'long', day: 'numeric', month: 'long' })
+  const date = new Date().toLocaleDateString(undefined, { weekday: 'long', day: 'numeric', month: 'short' })
   return (
     <div className="flex flex-col gap-4">
       <TabHeader
